requirements.txt
boto3==1.28.63
Flask==3.0.0
python-dotenv==1.0.0
gunicorn==21.2.0